Make content field nullable & improve code readability
This commit introduces several updates. In the database migration, the 'content' field on the 'projects' table has been made nullable. This flexibility was necessary as it was found that every project does not necessarily have content to add at the time of creation.

In 'ProjectResource.php', code structure has been improved for better readability. The 'image' upload has been moved to be closer to the form it's part of, and the 'content' field has been marked as nullable to reflect the database change.

Lastly, some minor changes have been made in 'welcome.blade.php' to improve the alignment and add Livewire styles and scripts to improve interactivity.